+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [0-based versioning](https://0ver.org/).
 ### Added
 
 - cppcheck for posix and libuv builds
+- Add `struct am_event_queue_policy::exclude` field
 
 ## v0.17.3 - 06-September-2026